feat(various): a bunch of new features!!
Introducing the new ScriptFileWriter which allows to write log entries to a file and persist the file afterwards in alfresco (or even mail it to another person)
Introducing the new ScriptJobService which allows to create new inline jobs, delete jobs and run the jobs. These jobs will be seen in the new scheduled job overview page of ootb-support-tools-addon

e.g.
jobs.scheduleTemporaryJob({
	jobName: 'SimpleAuthorModificationJob',
	runAs: 'system',
	cronExpression: '0 0/1 * 1/1 * ? *',
	script: function(){
		batchExecuter.processFolderRecursively({
			'root': repository.getCompanyHome(),
			onNode: function(node) {
				if (node.isDocument) {
					node.properties['cm:author'] = 'Alfresco';
					node.save();
				}
			}
		});
	}
});

Removal of the jmx related stuff because it does not work in community edition...

Introducing the new ScriptPermissionService which allows to set new permissions on nodes and also clear all permissions on nodes